chore: change return type of Count() from u64 to i64 (#92)

+++ datafusion/physical-expr/src/expressions/count.rs
@@ -24,7 +24,7 @@ use crate::{AggregateExpr, PhysicalExpr};
 use arrow::compute;
 use arrow::datatypes::DataType;
 use arrow::{
-    array::{ArrayRef, UInt64Array},
+    array::{ArrayRef, Int64Array},
     datatypes::Field,
 };
 use datafusion_common::Result;
@@ -96,7 +96,7 @@ impl AggregateExpr for Count {
 
 #[derive(Debug)]
 struct CountAccumulator {
-    count: u64,
+    count: i64,
 }
 
 impl CountAccumulator {
@@ -109,25 +109,25 @@ impl CountAccumulator {
 impl Accumulator for CountAccumulator {
     fn update_batch(&mut self, values: &[ArrayRef]) -> Result<()> {
         let array = &values[0];
-        self.count += (array.len() - array.data().null_count()) as u64;
+        self.count += (array.len() - array.data().null_count()) as i64;
         Ok(())
     }
 
     fn merge_batch(&mut self, states: &[ArrayRef]) -> Result<()> {
-        let counts = states[0].as_any().downcast_ref::<UInt64Array>().unwrap();
+        let counts = states[0].as_any().downcast_ref::<Int64Array>().unwrap();
         let delta = &compute::sum(counts);
         if let Some(d) = delta {
-            self.count += *d;
+            self.count += *d as i64;
         }
         Ok(())
     }
 
     fn state(&self) -> Result<Vec<ScalarValue>> {
-        Ok(vec![ScalarValue::UInt64(Some(self.count))])
+        Ok(vec![ScalarValue::Int64(Some(self.count))])
     }
 
     fn evaluate(&self) -> Result<ScalarValue> {
-        Ok(ScalarValue::UInt64(Some(self.count)))
+        Ok(ScalarValue::Int64(Some(self.count)))
     }
 }
 
@@ -148,8 +148,8 @@ mod tests {
             a,
             DataType::Int32,
             Count,
-            ScalarValue::from(5u64),
-            DataType::UInt64
+            ScalarValue::from(5i64),
+            DataType::Int64
         )
     }
 
@@ -167,8 +167,8 @@ mod tests {
             a,
             DataType::Int32,
             Count,
-            ScalarValue::from(3u64),
-            DataType::UInt64
+            ScalarValue::from(3i64),
+            DataType::Int64
         )
     }
 
@@ -181,8 +181,8 @@ mod tests {
             a,
             DataType::Boolean,
             Count,
-            ScalarValue::from(0u64),
-            DataType::UInt64
+            ScalarValue::from(0i64),
+            DataType::Int64
         )
     }
 
@@ -194,8 +194,8 @@ mod tests {
             a,
             DataType::Boolean,
             Count,
-            ScalarValue::from(0u64),
-            DataType::UInt64
+            ScalarValue::from(0i64),
+            DataType::Int64
         )
     }
 
@@ -207,8 +207,8 @@ mod tests {
             a,
             DataType::Utf8,
             Count,
-            ScalarValue::from(5u64),
-            DataType::UInt64
+            ScalarValue::from(5i64),
+            DataType::Int64
         )
     }
 
@@ -220,8 +220,8 @@ mod tests {
             a,
             DataType::LargeUtf8,
             Count,
-            ScalarValue::from(5u64),
-            DataType::UInt64
+            ScalarValue::from(5i64),
+            DataType::Int64
         )
     }
 }
